requirements.txt
Django==4.2
djangorestframework==3.14.0
gunicorn==20.2.0